+++ src/modules/dcc/broker.h
@@ -46,10 +46,10 @@ class KviDccBox;
 class KviDccZeroPortTag
 {
 public:
-	QDateTime    m_tTimestamp;
-	QString      m_szTag;
-	unsigned int m_uFileSize; // outgoing file size, valid only for file transfers obviously
-	unsigned int m_uResumePosition; // if 0 = no resume, valid only for file transfers obviously
+	QDateTime      m_tTimestamp;
+	QString        m_szTag;
+	unsigned long  m_uFileSize; // outgoing file size, valid only for file transfers obviously
+	unsigned long  m_uResumePosition; // if 0 = no resume, valid only for file transfers obviously
 };
 
 class KviDccBroker : public QObject
@@ -90,7 +90,7 @@ public:
 	void sendFileManage(KviDccDescriptor * dcc);
 
 	bool handleResumeAccepted(const char * filename,const char * port,const char * szZeroPortTag);
-	bool handleResumeRequest(KviDccRequest * dcc,const char * filename,const char * port,unsigned int filePos,const char * szZeroPortTag);
+	bool handleResumeRequest(KviDccRequest * dcc,const char * filename,const char * port,unsigned long filePos,const char * szZeroPortTag);
 
 public slots:
 	void rsendExecute(KviDccBox * box,KviDccDescriptor * dcc);
